+++ app/helpers/primer/form_helper.rb
@@ -7,5 +7,17 @@ module Primer
     def primer_form_with(**kwargs, &block)
       form_with(**kwargs, skip_default_ids: false, builder: Primer::Forms::Builder, &block)
     end
+
+    def primer_fields_for(record_name, record_object = nil, options = {}, &block)
+      fields_for(
+        record_name,
+        record_object,
+        options.merge(
+          skip_default_ids: false,
+          builder: Primer::Forms::Builder
+        ),
+        &block
+      )
+    end
   end
 end
